fix(messaging): make assembly scanning a setup extension
