+++ demo/Headless.Messaging.Aws.InMemory.Demo/Program.cs
@@ -3,10 +3,9 @@ using Headless.Messaging.Dashboard;
 
 var builder = WebApplication.CreateBuilder(args);
 
-builder.Services.ForMessagesFromAssembly(typeof(Program).Assembly);
-
 builder.Services.AddHeadlessMessaging(setup =>
 {
+    setup.ForMessagesFromAssembly(typeof(Program).Assembly);
     setup.UseInMemoryStorage();
     setup.UseAws(RegionEndpoint.CNNorthWest1);
     setup.UseDashboard(d => d.WithNoAuth());
