+++ demo/Headless.Messaging.RabbitMq.SqlServer.Demo/Program.cs
@@ -32,10 +32,9 @@ await using (var connection = new SqlConnection(AppDbContext.ConnectionString))
     );
 }
 
-builder.Services.ForMessagesFromAssembly(typeof(Program).Assembly);
-
 builder.Services.AddHeadlessMessaging(setup =>
 {
+    setup.ForMessagesFromAssembly(typeof(Program).Assembly);
     setup.UseEntityFramework<AppDbContext>();
     setup.UseRabbitMq("127.0.0.1");
     setup.UseDashboard(d => d.WithNoAuth());
