+++ src/Headless.Messaging.Core/IMessagingBuilder.cs
@@ -7,9 +7,9 @@ namespace Headless.Messaging;
 /// </summary>
 /// <remarks>
 /// <para>
-/// The messaging builder configures infrastructure and global behavior. Message consumers are registered
+/// The messaging builder configures infrastructure and global behavior. Explicit message consumers are registered
 /// directly on <see cref="Microsoft.Extensions.DependencyInjection.IServiceCollection"/> via
-/// <c>ForMessage&lt;TMessage&gt;(...)</c>.
+/// <c>ForMessage&lt;TMessage&gt;(...)</c>; assembly-scanned consumers are registered from the setup callback.
 /// </para>
 /// </remarks>
 [PublicAPI]
