+++ src/Headless.Messaging.InMemory/README.md
@@ -35,10 +35,9 @@ dotnet add package Headless.Messaging.InMemory
 ## Quick Start
 
 ```csharp
-builder.Services.ForMessagesFromAssemblyContaining<Program>();
-
 builder.Services.AddHeadlessMessaging(options =>
 {
+    options.ForMessagesFromAssemblyContaining<Program>();
     options.UseInMemoryStorage();
     options.UseInMemory();
 });
