+++ src/Headless.Messaging.InMemoryStorage/README.md
@@ -24,10 +24,9 @@ dotnet add package Headless.Messaging.InMemoryStorage
 ## Quick Start
 
 ```csharp
-builder.Services.ForMessagesFromAssemblyContaining<Program>();
-
 builder.Services.AddHeadlessMessaging(options =>
 {
+    options.ForMessagesFromAssemblyContaining<Program>();
     options.UseInMemoryStorage();
     options.UseRabbitMQ(config);
 });
