+++ src/Headless.Messaging.Pulsar/README.md
@@ -23,10 +23,9 @@ dotnet add package Headless.Messaging.Pulsar
 ## Quick Start
 
 ```csharp
-builder.Services.ForMessagesFromAssemblyContaining<Program>();
-
 builder.Services.AddHeadlessMessaging(options =>
 {
+    options.ForMessagesFromAssemblyContaining<Program>();
     options.UsePostgreSql("connection_string");
 
     options.UsePulsar(pulsar =>
