+++ src/Headless.Messaging.Redis/README.md
@@ -29,10 +29,9 @@ dotnet add package Headless.Messaging.Redis
 ## Quick Start
 
 ```csharp
-builder.Services.ForMessagesFromAssemblyContaining<Program>();
-
 builder.Services.AddHeadlessMessaging(options =>
 {
+    options.ForMessagesFromAssemblyContaining<Program>();
     options.UsePostgreSql("connection_string");
 
     // Queue delivery through Redis Streams.
